readme: updates readme with new team information
This is a start for what our new README should contain. As our project
becomes more and more mature, we should change the README. Right now
there is a ton of stuff left behind from the old group in our repo that
may not be particuarly relevant to what this new group is doing.