front-end/localStorage/userOperations: fix cors failing on login
There is a bug where fetch requests aren't accepting session cookies
from the backend. This is because of the credentials: 'same-site' option
that is being used in the request. This fixes the issue by changing the
option to 'include'. This isn't as strict as it maybe should be, but it
isn't neccisarily insecure. Most of the security is stored in the cookie
headers which is set in the backend. In a theoretical case where an
attacker was injecting JS into the page, they would still be able to
send those cookies with an injected fetch request that used 'include' as
the credentials option. Since what were shipping is our own trusted
code though, this theoretical scenario isn't anything we should worry
about.