back-end/index: fixes cookies not working in production due to proxies
The back-end is behind a reverse proxy and communicates to that proxy
via HTTP. Secure cookies will not work over HTTP, so despite the
reverse proxy only sending traffic over HTTPS, express refused to send
the cookie. This fixes the issue by having the reverse proxy set the
X-Forward-Proto header (not something handled in this commit or repo),
and setting the express-session value 'proxy' to true when in
production.