back-end/test/readFiles: add tests for readFiles
Tests the readUser function. This should be considered a bit of an
integration test because it relies on updateUser working correctly. This
is because of encryption. It is not possible to insert directly into
Mongoose without encrypting everything by hand. Eventually this process
becomes more error prone than just trusting that updateUser is doing the
right thing.